Added max-results param for Spreadsheets List Feed
While working with Spreadsheets list feed, insertRow method loads the
whole feed for selected worksheet to get the edit URL. When the
worksheet has a lot of records this may take a while and even lead to a
crash. That's where max-results parameter comes in handy. It limits the
number of rows returned by list feed query. It can also be zero so it
returns only a header of the list feed.